requirements.txt
PyMuPDF==1.22.0
pytesseract==0.3.10
Pillow==9.5.0
numpy==1.24.3

